+++ changelog.md
@@ -10,6 +10,7 @@
 
 - Drop support for Python 3.8 to follow Home Assistant changes
 - Return `bytes` from to_knx() in knxip package instead of `list[int]`
+- Add a callback for `connection_lost` of TCP transports to Tunnel
 
 ## 0.19.2 TCP Heartbeat 2022-02-06